refact: user 관련 레이어 위치 정리 및 userRepository 인터페이스 생성과 해당 인터페이스 구현
- 도메인을 기준으로 레이어 파일 모으기
- 테스트 용이성을 위해 userService 에 사용되는 여러 레포지토리를 하나의 레포지토리에 구현

+++ src/main/java/study/back/dto/item/UserManagementInfo.java
@@ -13,6 +13,6 @@ public class UserManagementInfo {
     private Long id;
     private String email;
     private String datetime;
-    private int point;
-    private int commentCount;
+    private Long point;
+    private Long commentCount;
 }
