+++ src/main/java/study/back/dto/response/GetUserResponseDto.java
@@ -6,15 +6,15 @@ import lombok.NoArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import study.back.dto.item.UserItem;
-import study.back.entity.UserEntity;
+import study.back.user.entity.UserEntity;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class GetUserResponseDto extends ResponseDto {
     private UserItem user;
-    private int totalPoint;
+    private Long totalPoint;
 
-    public static ResponseEntity<GetUserResponseDto> success(UserEntity user, int totalPoint) {
+    public static ResponseEntity<GetUserResponseDto> success(UserEntity user, Long totalPoint) {
         GetUserResponseDto responseBody = new GetUserResponseDto();
         responseBody.code = "SU";
         responseBody.message = "유저 정보 받기 성공";
